requirements.txt
# Core
gradio>=5.0,<6.0
soundfile>=0.12
numpy>=1.26
librosa>=0.10  # time-stretch post-process (ralentir l'output sans changer le pitch)

# TTS — Chatterbox Multilingual (Resemble AI), SOTA open-source 2026
# MIT license, 23 langues dont le FR, zero-shot voice cloning
chatterbox-tts

# Chatterbox pin torch==2.6.0 mais ne touche pas torchvision, ce qui casse
# l'ABI sur Colab (erreur "operator torchvision::nms does not exist").
# On force torchvision à la version compatible avec torch 2.6.0.
torchvision==0.21.0

# Testing (dev only, mais pas grave de l'avoir dans Colab)
pytest>=8.0
